Coq: Define expressive and expresses
_Expressive_ wallets are more restricted than previously:
All amounts below the maximum (sum of all coins) can be expressed via a
subset of coins. This includes amounts above the target.
A wallet _expresses_ a target if the wallet is _expressive_ and if the
wallet sums to at least the target.
The correctness proof for next_coin needs to be updated.